generator.cのソースを複数ファイルに分割する
