+++ src/generator/register.c
@@ -1,0 +1,35 @@
+static char *reg32[9] = {
+    "edi",
+    "esi",
+    "edx",
+    "ecx",
+    "r8d",
+    "r9d",
+    "eax",
+    "ebp",
+    "esp",
+};
+
+static char *reg64[9] = {
+    "rdi",
+    "rsi",
+    "rdx",
+    "rcx",
+    "r8",
+    "r9",
+    "rax",
+    "rbp",
+    "rsp",
+};
+
+static char *reg8[9] = {
+    "dil",
+    "sil",
+    "dl",
+    "cl",
+    "r8b",
+    "r9b",
+    "al",
+    "epl",
+    "spl",
+};
